pyproject.toml
[project]
name = "devfest-cerrado-genai"
version = "0.1.0"
description = "Desvendando a GenAI: Da Teoria às Aplicações com Gemini e Gemma - DevFest Cerrado 2024"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Vinicius Caridá" }
]
dependencies = [
    "google-generativeai>=0.8.0",
    "python-dotenv>=1.0.0",
    "jupyter>=1.0.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "matplotlib>=3.8.0",
    "Pillow>=10.2.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.14.0",
    "black>=24.0.0",
    "flake8>=7.0.0",
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]

[tool.black]
line-length = 100
target-version = ['py310']
